Refactor
Place each schema and router in its own file.
